Add files via upload
In this assignment, you will load the various tables from the Rebrickable website and perform analysis on the data. Rebrickable is a fantastic website that is used to show you which LEGO sets you can build from the sets and parts you already own. They will even show you which parts you are missing to complete the new set and suggest sets that you could purchase to get those parts.